field for inputting collection
