+++ Controllers/SortingController.cs
@@ -1,7 +1,18 @@
+using Microsoft.AspNetCore.Mvc;
+
 namespace sortingAlgos.Controllers
 {
-    public class SortingController
-    {
-        
-    }
+  public class SortingController
+  {
+    // [HttpGet]
+    // public IEnumerable<int[]> Get(int[] collection)
+    // {
+    //   var collectionItterations = Sorter.BubbleSort(collection);
+    //   // return Enumerable.Range(1, 5).Select(index => new WeatherForecast
+    //   // {
+    //     // collection = collectionItterations;
+    //   // }
+    //   return collectionItterations;
+    // }
+  }
 }
